requirements.txt
python-telegram-bot>=2.4 
requests>=2.32.3 
langdetect>=1.0.9 
python-dotenv>=1.0.1